WIP Register different action and resource name for v2.1 on v3
This patch adds an APIRouterV21 class and the api-paste code
required to implement a /v2.1/ REST API resource.

This patch uses a decorator to implement a different action
name for migrate live for V2.1 (keeping with V2) compared to
the V3 API.

The infrastructure is also added to supply an optional version
number to the Keypairs plugin get_resources method which allows
it to return a different resource name for V2.1 compared to V3.

Partially implements blueprint v2-on-v3-api

Change-Id: I8476885e6892632e2fdc4af6923d2d3f25aa1713