Adds separate plugin namespace for V2.1 with V3 framework
Ads a separate plugin namespace in setup.cfg for the V2.1 API
in setup.cfg. This will allow the V2.1 API to have a different set
of plugins loaded in the same deployment than for V3. This is necessary
so that the V3 API is not forced to load plugins such as images
or volumes which the V2.1 needs to support for backwards compatibility
reasons

Partially implements blueprint v2-on-v3-api

Change-Id: I61f5c818b900e085966a198ca518b98862e8fdd9