pick parameters for evacuate
This patch decouple the view and data.

For parse json-schema to fetch parameter from request body.
This patch only implement for input. The implementation of
output need output's json-schema. The nova api python code
just need provide a dict of all the parameter. Then parse
jsonschema to generate the output.

@schema:evacuate:3.0

Change-Id: I43f93b968200b06e1c3c570c3348471192ed6b29