[POC] micro-version: Add multiple version support
This patch add multiple version support for wsgi.

The wsgi will get request version from header "X-OS-Compute-Version"
If the "X-OS-Compute-Version" isn't specified, the OS_COMPUTE_MIN_VERSION
will be used. The value of OS_COMPUTE_MIN_VERSION is "2.1".

The decorator version is used for api method to indicate the version
supported.

decorator "version" accept two parameter, one is for base version, another
is for max version the method supported. If the request version is out of
range of base and max, wsgi will return HTTPNotFound.

This patch use evacuate extension as example.

TODO
  * The api method without decorator version means support all the version.
    This isn't expected behavior. We should add decorator version to all the
    api method.

Change-Id: I20a9b0b01fdd5cd1dd593edea37fc3fc5820f82d