libvirt: Enable driver configures PMEM namespace when initiate
Add three configure options for configuring the namespace.

CONF.libvirt.pmem_region_names: The region can be used for nova.
CONF.libvirt.pmem_namespace_sizes: The size for each namespace,
also the number of sizes in the list will be the number of namespace
want to create.
CONF.libvirt.pmem_namespace_prefix: the prefix is used to build
the name for each namespace.

Change-Id: Ia6391d61932acc1f9ef8751be322e5143d9facf5