libvirt: report pmem resources by provider tree
Reports pmem namespaces to the provider tree. Since the namespace
is created when startup, and each namespace will be the minimal allocatable
unit. So each namespace will be a resource provider, and the total,
min_size, max_size and step_size are equal to the size of namespace,
then this make sure the whole namespace will be allocated.

Change-Id: I77590e41a3667c4b22126147b12f3ea14dc7fbb9